finalize graphic button element
