+++ src/components/GraphicButton.js
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import { motion } from 'framer-motion';
 import cx from 'classnames';
 import create from '../assets/images/create-button.svg';
-import open from '../assets/images/open-button.svg';
 
 const GraphicButton = ({
   children,
