+++ stories/GraphicButton.stories.js
@@ -2,15 +2,14 @@ import React from 'react';
 import { action } from '@storybook/addon-actions';
 import Harness from './helpers/Harness';
 import GraphicButton from '../src/components/GraphicButton';
+import openIcon from '../src/assets/images/open-button.svg';
+import colors from './helpers/Colors';
 
 import '../src/styles/_all.scss';
 
 const requiredProps = {
-  // name: faker.lorem.sentence(),
-  // schemaVersion: 5,
-  // lastModified: faker.date.recent().toISOString(),
-  // installationDate: faker.date.recent().toISOString(),
-  // description: faker.lorem.sentence(),
+  children: <><h2>Action</h2><h3>detail text</h3></>,
+  onClick: () => action('onClick')
 };
 
 export default { title: 'GraphicButton' };
@@ -18,11 +17,42 @@ export default { title: 'GraphicButton' };
 export const normal = () => (
   <Harness
     requiredProps={requiredProps}
-    graphicPosition="-4.5rem bottom"
-    color="slate-blue--dark"
-    graphicSize="auto 105%"
-    onClick={action('onClick')}
   >
-    {props => <GraphicButton {...props}><h2>Open</h2><h3>from Computer</h3></GraphicButton>}
+    {props => <GraphicButton {...props} />}
   </Harness>
 );
+
+export const userIcon = () => (
+  <Harness
+    requiredProps={requiredProps}
+    graphic={openIcon}
+  >
+    {props => <GraphicButton {...props} />}
+  </Harness>
+);
+
+export const colorVariants = () => {
+  return colors.map(color => (
+    <Harness
+      requiredProps={requiredProps}
+      color={color}
+    >
+      {props => <GraphicButton {...props} />}
+    </Harness>
+  ));
+};
+
+
+
+// export const userIcon = () => (
+//   <Harness
+//     requiredProps={requiredProps}
+//     graphicPosition="-4.5rem bottom"
+//     color="slate-blue--dark"
+//     graphic={openIcon}
+//     graphicSize="auto 105%"
+
+//   >
+//     {props => <GraphicButton {...props}><h2>Open</h2><h3>from Computer</h3></GraphicButton>}
+//   </Harness>
+// );
