+++ stories/helpers/Colors.js
@@ -1,0 +1,35 @@
+const colors = [
+    'neon-coral',
+    'neon-coral--dark',
+    'mustard',
+    'mustard--dark',
+    'sea-green',
+    'sea-green--dark',
+    'white',
+    'slate-blue',
+    'slate-blue--dark',
+    'navy-taupe',
+    'cyber-grape',
+    'rich-black',
+    'charcoal',
+    'platinum',
+    'platinum--dark',
+    'sea-serpent',
+    'sea-serpent--dark',
+    'purple-pizazz',
+    'purple-pizazz--dark',
+    'paradise-pink',
+    'paradise-pink--dark',
+    'cerulean-blue',
+    'cerulean-blue--dark',
+    'kiwi',
+    'kiwi--dark',
+    'neon-carrot',
+    'neon-carrot--dark',
+    'barbie-pink',
+    'barbie-pink--dark',
+    'tomato',
+    'tomato--dark',
+  ];
+
+export default colors;
